feat: add gwatchdog package
Integrate it with the top level of the Engine and with the state
machine. Next week will start integration with the mirror and the
components that live outside the engine.

The watchdog asserts that the main loop of any consumer remains
responsive. If a loop gets stuck and is unable to respond to the
watchdog's signal within a configured timeout, the watchdog cancels the
root context. A consumer should check if the root context was a watchdog
termination (using the gwatchdog.IsTermination function), and if so,
produce more detailed logs upon shutdown than it would otherwise.

This is the first pass at the API and it is likely to evolve more over
the next few commits.

The Watchdog type may change from an struct to an interface. The
watchdog may provide a sink for termination logs so that termination
logs go to a more permanent and discoverable location than stderr.

The signal sent to consumers will likely get more fields, including a
state dump request, so that an operator can get a meaningful
introspection on the full state of the system.

The way the components provide a MonitorConfig may change too, as the
operator may need some control over the timing values there.