refactor: state machine accepts StateMachineRoundView
Accepting a VersionedRoundView was a fine starting point, but now there
are multiple signals that the mirror may send to the state machine. The
current round view is the obvious one, but the mirror may also need to
signal that it is jumping ahead at least one round. Lastly, if the state
machine has fallen behind, the mirror may need to signal an entire
committed block for the state machine to immediately finish the height.