feat: teach StateMachineViewManager about jump ahead rounds
The mirror is aware of when a round is skipped over due to a later round
having enough prevotes or precommits. Now the mirror can send a signal
to the state machine indicating that the mirror is jumping past a round
despite not having sufficient precommit info for the jumped-over round.

The state machine does not yet interpret this signal.