feat: add initial gordian-stress command
It doesn't do much yet. It is able to start a libp2p seed host, and it
provides a socket file that reports the libp2p seed's addresses. This is
to be used for bootstrapping a network.

This is the first step towards a robust toolset to simplify starting
networks of Gordian instances. The existing echo server commands are
also simple, but they lack flexibility, for example to configure unique
settings on each node in the network.