feat: add engine metrics
These are split into state machine and mirror metrics. We are explicit
in that this should not be considered a stable API. This can be used for
general monitoring, but it is going to be more immediately used as a key
feature in halt reproduction with running several instances of
gordian-stress concurrently.