feat: send halt signal to validators on validator drift
Now that the seed is observing engine metrics from connected validators,
it occasionally checks the metrics to discover if any validator has
either drifted from the rest of the validators, or if a state machine
has diverged from its mirror. If any of those conditions happens, then
the seed sends a halt signal to all its validators, causing a quit with
verbose output.