fix: use slog in gcosmos
Trying to derive an slog logger from zerolog, even when using an
off-the-shelf third party library, was causing slog messages to be
dropped. I'm not sure why. So instead, build our own minimal wrapper
around slog to satisfy the cosmos logger interface, and overwrite the
command's logger to use that. It is quite possible other parts of the
Cosmos stack may fail due to expecting zero log, but since only this
command calls into Gordian code, it might be okay.