test: add initial CLI test
This just gets the skeleton in place, making it trivial to run a
sequence of commands without having to fall back to the shell or to
manually manage multiple config directories.

It's not very close to complete but it is a good starting point to
continue expanding the genesis, which will then allow Gordian to agree
with simapp about the validator set.