chore: add initial gserver.Component sketch
This was compiling and somewhat running in another branch, but now we
have cleaned it up a little and ensured it compiles against the current
version of the SDK we are using.

The next steps are to advance the SDK to the head commit and handle
patching the SDK to inject the Gordian server component instead of the
CometBFT component.