test: increase stake past magic value
When toggling the flag to start with Comet instead of Gordian, the
handshake appears to never complete with a small amount of stake in the
genesis accounts. At this updated value, if you also use 1 validator
instead of 4 and run with Comet, you can start producing blocks.

This isn't necessarily directly helpful for Gordian, but it does give a
useful reference point for future issues where Comet proceeds farther
than Gordian.