fix: respect InitChain validators in mirror kernel
This was an outstanding TODO comment, that we had not touched in any
existing tests.

The typical SDK flow at genesis is that there are no validators reported
to the consensus layer, until InitChain completes. As a result, the
mirror was panicking because it thought there were no validators. Now it
should respect the results of InitChain.