feat: read signing key from comet config
This allows us to continue hooking in to the existing simapp
infrastructure that expects comet everything. Since we now are running
the single validator and we provide its signing key, Gordian is able to
advance rounds in the first height.

Next up is figuring out how to get proposed block details from the SDK,
which is probably going to involve something with a mempool.