test: allow starting with multiple validators
The current version of the SDK appears to have bugs for validator counts
above 1 and below 11. From 2 through 9 inclusive, it only reports half
of the validators in the validator updates from init chain. At 10, there
is a storage-layer panic. For whatever reason, 11 and above appears to
work correctly.

So from here, we need to add a new stake strategy to allow us to start
11 validators, where the first 4 validators comprise >2/3 of the total
delegations. If the remaining validators are offline, that won't affect
the supermajority.

And we need a seed node so the libp2p layer can communicate, or we need
to capture the first validator's address and share it with the
subsequent validators.