feat: add seed subcommand to start libp2p seed
And teach the Gordian-flavored start command how to accept a seed
address and connect to it at startup.

With this in place, it appears that each of the 4 active validators in
the multiple validator test see one prevote at height 1 / round 0, but
with the dynamic keys, it is unclear whether it is always the same
validator or a different one each time.

Since the validators appear to be communicating, the next piece to fix
is the gcosmos consensus strategy, which currently assumes every
validator is due to propose every round.