fix: only allow one proposer at a time
This was an accidental carryover from only running a single validator.
Every validator was trying to propose its block, and it was proposing
its block during the libp2p settle time, so nobody saw anyone else's
proposed blocks or votes.

With this change, and with increasing some test timeouts, we see rounds
advancing within height 1, which indicates that the validators are
sharing nil votes properly. I do not yet know why the proposed blocks
appear to be getting dropped, but that will be the next point of
investigation.