feat: add initial sketch of TxBuffer
This is currently a "validator-local mempool" that only works in a
single validator configuration.

To prototype this, we just pass around a TxBuffer and associated mutex;
but even in sketching this out, it was trivial to end up with a
deadlock, hence the inlined, narrowly scoped, locking functions.

This subsystem will also need a central coordinating goroutine. There
are only three lock calls, so that is a small area of code that may need
to change.

The current test asserts that the fixed account has the expected
starting balance, and that it drops by 100 following the transfer (and
the receiver increases by 100 as well).

There are some oddities in the current test -- namely, we are using
bogus values for the account number and the sequence, yet simapp is
accepting those values still. I don't follow why these are not being
validated, but maybe that is just an expected quirk of simapp.

Lastly, this also includes another patch to the SDK to allow us to
simulate multiple pending transactions building upon each other.