test: sketch multiple validator transaction test
Add most of the structure for the test, but skip the test for now as it
is not yet complete. It is missing more assertions at the end.

And add a /debug/pending_txs endpoint. It should be easier to just
inspect those, rather than trying to intelligently guess which validator
a transaction should be sent to.