fix: use block annotation for timestamp when delivering block
The prior call to time.Now() led to a divergence of app state hash
across multiple validators, unsurprisingly.

This does not have a dedicated test yet. There is at least one missing
check in the state machine, and that is where we will add the test to do
an early rejection of proposed blocks that have a differing app state
hash.