fix: ignore proposed blocks that mismatch the state machine
If we receive live updates of proposed blocks that have a different set
of validators or a different app hash from what we have, silently ignore
the proposed block.

It would probably be more informative to at least debug log the ignored
block, but that would entail repeatedly logging out the same ignored
proposed blocks multiple times within a round, or we would need to track
a bit more state to ensure we only log a proposed block once. So, for
now, it is silently ignored.