feat: use updated app data ID
This app data ID is part of the proposal, structured as:

HEIGHT:ROUND:NUM_TXS:TXS_HASH

Where these parts indicate that the proposer is on the agreed height and
round; how many trnasactions the other validators should expect in the
transaction set; and the hash of concatenating the ordered set of
transactions' hashes.

As of this commit, the consensus strategy is hardcoded to always
indicate zero transactions. The next step is to teach the consensus
strategy how to produce a correct app data ID for a non-zero set of
transactions. That will currently cause the other validators to panic,
so after that, the proposer needs to somehow provide the app data and
the other validators need to receive or retrieve it.