feat: rough sketch to start proposing with pending transactions
This has a lot of issues, but it runs correctly for a single validator
network.

This highlights a current design issue. Right now, when the mirror
receives a proposed block, the mirror first does some basic validation
(such as verifying the signature of the proposed block and the hash of
the underlying block), and then the mirror passes the proposed block to
the state machine. The state machine always keeps the full list of
proposed blocks for the current height and round, but when it sends the
proposed blocks to the user-defined consensus strategy, it only includes
those which match the current app hash and calculated validator set
(with the purpose of not allowing the driver to consider proposed blocks
from a different validator with divergent state).

The problem is that now, when we receive a proposed block with an app
data ID, we have to assume that we depend on an asynchronous event to
retrieve the data behind the ID. That may be listening on a network
connection, or it may be opening an active connection to a remote host
somewhere.

The simplest solution may be to include an extra channel in the
arguments passed to ConsiderProposedBlocks, or maybe to
to ConsiderProposedBlocks, or maybe . This would allow a flow like:

1. The state machine sees a proposed block and passes it to the
   consensus strategy.
2. The consensus strategy recognizes that it does not yet have the data
   backing the app data ID, so it either starts asynchronous work to
   retrieve it, or it has to otherwise passively wait for the data to
   arrive by other means.
3. The call to ConsiderProposedBlocks completes, and afterwards the data
   arrives by whatever means.
4. The state machine needs to tell the consensus strategy to re-consider
   the proposed blocks now that new app data is available.

With only that mechanism in place, we should be able to handle either
case of proactively fetching the app data or passively waiting for it to
arrive, without requiring more active work within the state machine or
its subsystems.

Note, however, that implementing this does not address any of the
necessary cleanup to the pending transaction buffer. That still needs to
be adjusted to use the central goroutine pattern, and it needs expanded
to have finer control over removing transactions that arrived in a
different proposed block, and possibly splitting the buffer when the
current transactions being included into a proposed block.