refactor: include reason for considering proposed blocks
Add a new reason argument to the ConsiderProposedBlocks method. This
currently indicates which proposed blocks' hashes are new compared to
the previous call; but more importantly, it provides the structure for
re-consideration following app data retrieval. The consensus engine does
not involve itself in retrieving app data associated with a particular
block, so we assume the driver will initiate some background work,
preventing the first observation of a proposed block from being
accepted. The driver or consensus strategy may then begin work to
retrieve that data. Once it is retrieved, the
ConsiderProposedBlocksReason.UpdatedAppData field will indicate which
app data IDs the consensus strategy may assume are now available.