feat: state machine responds to block data arrival
When block data arrives and a message is sent on the BlockDataArrival
channel, the state machine causes a new call to ConsiderProposedBlocks,
indicating which new block data have arrived.

If the block data arrives before the proposed block, it is the consensus
strategy's responsibility to check for block data when the consider
reason indicates a new proposed block. But, that would only happen in
configurations where it is actually possible to receive app data before
the proposed block. If the consensus strategy sees a new proposed block
and then starts background work to retrieve the block data, then it
would not be possible to get block data first.