chore: rename AppDataID to just DataID
We are trying to get away from the "app" word in general. We probably
won't be able to eliminate it everywhere. But DataID should be clear
enough in the context of a block.

We still use "block data ID" scoped to the driver in gcosmos, as its gsi
package has a bit broader context where simply "data ID" is not clear
enough on its own.