feat: add block data hosting capability
The simple round trip test for this is failing -- it is currently
skipped, because the failure is more of a test issue on correctly
encoding and decoding the set of transactions. The serialized data is
transferred correctly, there is just a mismatch on serializing and
deserializing the data. We'll get that sorted out shortly.