feat: retrieve remote block data
The other validators are retrieving the block data over libp2p from the
proposer. However, there is currently an app hash issue: the proposer's
app hash differs from the non-proposing validators, when running the
multi-validator transaction test (which is still skipped for now).

It looks like this is an issue where the driver is pulling from the
pending transaction buffer when it is time to finalize a block. The
proposer has populated the pending transaction buffer, but the other
validators have only simulated the transactions without modifying the
pending transaction buffer.

This appears to highlight a need for a separate mechanism to temporarily
store the transactions associated with a particular block.