fix: synchronize on (*gdatapool.Pool).EnterRound
This ultimately addresses a data race where EnterRound on the pool was
unintentionally running concurrently with a call to Need or
SetAvailable, and EnterRound was clearing the current request map at the
same time that a new entry was being added to the map.

So, in order to ensure that the clear operation finishes before any
writes, EnterRound needs to respond that it is finished. That also
highlighted that the consensus manager had a potential data race, in
that the other input channels were still buffered. This could lead to a
case where a different input channel had a buffered input while
EnterRound was happening, and the select statement could randomly choose
to handle EnterRound before the other input. Then, we would handle an
old input for a new round, potentially putting the Consensus Strategy in
an invalid state.

Switching those channels to be unbuffered then caused some state machine
tests to fail. Those failures were temporarily mitigated with a timeout
on the send instead of expecting a non-blocking send.

The longer term solution for this, is to have the state machine
gracefully handle blocked sends to the consensus manager, accumulating
its inputs in a buffer. I think we would have needed to do this
eventually anyway; at worst we just have to do it sooner than otherwise.