chore: update to go1.23 and latest SDK
go1.23 was released yesterday. The main branch of the SDK has already
marked go.mod as using 1.23, so we will go ahead and match it here too.

I have been running with go tip for a while and everything seems to run
fine, so I'm not worried about runtime changes breaking anything here.