feat: respect positive delegations
More specifically, respect the updated powers in the validator updates
reported after delivering a block.

The added test currently only covers an increased validator power. We
will need to see what happens when a validator is jailed or otherwise
removed from the active pool. I suspect it is just an update with a
power of zero, but we need to confirm that with a test.