chore: update to latest SDK
The patches got rearranged slightly here, as one old patch was
invalidated due to a matching upstream change. Then, we had to add a new
patch at the end to deal with the --home argument being set upstream
directly against the viper value, which meant that the --home flag was
ignored again in some instances.

Also attempt to set .gitattributes for the patch files so that the
whitespace stops changing upon commit.