feat: add gassert package
Since we are about to start on state machine catchup, I am adding the
assertion environment to the mirror and state machine.

Previous debugging in those core subsystems has been tedious, and the
effort placed there has not paid forward, although of course it helped
with whatever the bug at that time was.

The idea of the assertion environment is that any subsystem can have a
set of runtime invariants that can be checked via a two-stage opt-in
system -- both setting the "debug" build tag and enabling the rule
specifically or with a wildcard.

I did not have any particular assertions fresh in my mind when adding
this, but I expect we will add the first one or two while finishing the
state machine catchup implementation.