chore: add FinalizedBlockHash field to RoundLifecycle
For state machine catchup, we also need to call rlc.CycleFinalization,
but we will not have a PrevVRV from which we can determine the most
precommitted hash. So instead, we just trust what the driver reports
back as the finalized block hash.

Add first invariant check on RoundLifecycle, ensuring that we always
have a finalized block hash available. Only one test needed to be
updated in response to that check.