fix: handle state machine catchup on round advancement
We already handled it when advancing height; refactor the advanceHeight
and advanceRound code to share the common part, in order to properly
enter catchup during any advancement. It's probably unlikely to be
triggered unless there is a considerable slowdown while finalizing the
block, but we certainly cannot rule out the possibility of it happening.