refactor: use ValidatorSet in Block instead of slice of Validator
We often want to couple the ordered collection of validators with the
hashes representing that collection, and the ValidatorSet serves that
purpose. This was a relatively narrow change that touched a wide part of
the codebase.

This change was long overdue. We recalculated the validator hashes in
many spots, and now we calculate them in few spots.

There is only one failing test (the tmintegration validator shuffle
test), and it is skipped for now. I'm not yet sure if it's a bug in that
test or if we are not correctly recalculating a hash when only the order
of the existing validators changes. I observed that the test does not
receive a finalization request from the engine, so that will be the
starting point for the investigation next week.

I believe all the updated fields referencing a ValidatorSet have been
renamed from XValidators to XValidatorSet, but it is possible I missed
one or two along the way.

Also, this is a move back towards preferring byte slices that are not
intended to be modified, rather than strings that are more obviously
immutable. After experimenting with the string approach for a while
here, byte slices feel more appropriate. They are much more heavily used
in the internals, where we can trust the core maintainers to use them
correctly. In the future we will have more runtime assertions to prove
they work correctly, and sometimes the data race detector would likely
pick up on unsynchronized mutation of them too.