refactor: rename tmconsensus.Block to Header
Header is more in line with common terminology in the blockchain space.
This allows us to distinguish a Block as (Header + Data), so a Header
implies that Data is not necessarily available.

The consensus strategy still uses Proposed Blocks in its method names.
Even though we only provide it the headers directly, there is the
out-of-band mechanism to get the proposed blocks' data, and we assume
that data is necessary for the consensus strategy to make its decision
about the proposed blocks.

Most variable names, and a lot of documentation and comments, have also
been updated to reflect the new name; but I've surely missed many spots.