chore: update to latest SDK
TxResult was moved from appmanager to core/server, which was mostly a
seamless port, but it did touch the result of the gRPC response for
transaction submission.

I am not sure which SDK commit introduced it, but setting the height on
the init chain request to zero instead of one is now a requirement;
without that, only a single validator is accepted, for some reason.