feat: initial replay handling in the mirror
There are some missing edge cases and associated missing tests, but this
is (most of) the basic functionality for the happy path.

Now the mirror is able to emit a lag signal, and the driver has a
channel where it can send the headers it has fetched. How the driver
fetches those blocks is completely independent of the engine internals.