chore: add but do not enable HeaderSyncClient
This is almost working in integration with gcosmos, but there is a
difference in the way the fixtures work compared to the integration
tests, causing a panic.

It looks like, in fixtures, the initial height has a PrevCommitProof
corresponding with the pseudo-genesis block. But we apparently are not
setting the PrevCommitProof field at the initial height in a typical
workflow, as the field is nil in both tmintegration and gcosmos.

I need to take a closer look to understand if we can have the field
populated. But there is a lot of work in this commit that is near enough
to completion that it makes sense to commit and push it now, to avoid
mentally juggling those changes and the separate proof debugging work.