feat: add PrevCommitProof to RoundView
This is a significant part to general bookkeeping that isn't yet fully
wired. This commit adds the previous commit proof as part of the round
view, to enable validating it everywhere necessary, but this is not yet
complete. The next step is to confirm it is correct on incoming
messages and to update tests accordingly.