feat: properly set previous commit proof
This was an old outstanding TODO, which was not an issue until we were
doing validation of replayed blocks fetched from untrusted sources.

The tests for this were fairly straightforward, aside from hitting an
effective off-by-one bug when setting the previous commit proofs while
validator sets were being changed; so the new longer tests are asserting
correct behavior of previous commit proofs during validator changes.

This should unblock mirror catchup now that we can fully validate
replayed blocks by verifying their previous commit proofs.

There are some outstanding TODOs in this commit which also need to be
addressed, but for now this is functional enough to produce blocks.