feat: enable header sync client
This was one key piece to mirror catchup, disabled until we got all the
previous commit proof wiring corrected. Now, a late-arriving validator
is able to sync from its connected peers to reach the current voting
height. However, it will fail if there are any transactions in any
blocks it syncs, so that is the next feature to implement.