refactor: encoded raw data length in data ID
I almost moved this from the proposal annotation to a block annotation,
but it occurred to me that you would still need the block annotation --
not just the data ID -- to validate the decoded size when fetching the
block data in that case.

Now this is an unambiguous indication of the decoded data size, which
makes it easier for clients to safely decode data. Moreover, if the
block is committed and the validator signatures are verified, that is
proof that the decoded data size is correct and trustable.

Also update some other comments in the datapool package while reading
through it and trying to decide how and if it should be integrated with
block data retrieval.