feat: add block data to gp2papi
In addition to the existing /gcosmos/committed_headers/v1/height path,
we now have a /gcosmos/full_blocks/v1/height path. We are likely to keep
the former path for "mirror only" configurations. The latter path is
more useful for proper mirror catchup when a state machine is involved.

Next, we are going to have to update the HeaderSyncClient type to pull
from the new endpoint by default, and to do something meaningful with
the retrieved block data.