feat: read from full block endpoint in SyncClient
The SyncClient is now using the correct endpoint, which passes the
current tests, but would fail if it tried to replay a block containing
data (which would have already failed on the prior commit, albeit in a
different way).

Next is updating the name of the ReplayedHeaderRequest/Response types,
and teaching the SyncClient how to decode the block data. This work
could arguably be pushed into another part of the driver, but ultimately
this goroutine is going to be blocked on that work one way or another,
so it may as well do the work itself.