refactor: extract block data encoder and decoder
Use them in the libp2p host and client for block data.

The new encoder and decoder use uncompressed data when the snappy
encoding results in a larger size than the plain uncompressed data.

I am not happy with the framing format of the encoded block data, which
is currently a JSON-marshalled [][]byte. That ends up being encoded as a
JSON array of strings that are themselves encoded in base64. But, while
we are stuck with JSON, that is probably fine.

Also add a HashOnlyTransaction type for simple use within internal tests
where we need to encode or decode transactions.