feat: enable excluding peers in sync client
There are many error cases where we can have received data with invalid
formatting. In those cases, we exclude the peer from any future sync
requests.

We may want to switch this from a map to some kind of LRU or timeout
cache, but a map is fine for this stage.