feat: add gsbd.RequestCache
This is a side channel for the driver to plumb block data from before
interacting with the engine, through to the driver-side when finalizing
a block.

The driver was already expecting the block data to be available during
normal consensus, but we had no way to get block data to the driver
during mirror catchup. I think we will need to include the request cache
in the normal consensus flow, so that the driver only has to look in one
place to discover block data.

The request cache includes the encoded transaction data that we have
retrieved with the header, so that it can directly store that in its
BlockDataStore, hosting it for other peers as well.