chore: rename sync client to catchup client
This client is responsible for dealing with lagging mirror state and
catching up. The word "sync" was not specific enough.

It was more important to distinguish this behavior from an upcoming
different client for retrieving block data from a proposer (which is
somewhat already in the datapool set of types, but which is not as well
defined as I would prefer it to be).