feat: add gsi.PBDRetriever type
This is a generalized client to fetch proposed block data. The only
specialization it has so far is libp2p, which is also the only transport
we provide so far.

This only adds the code with a test, but does not integrate it into the
driver yet. There is a bit more code (libp2p specifics) that needs to be
removed from the gsbd package in favor of this type.