fix: handle first replayed header having a non-zero round
There was a bug in the kernel state such that a replayed header for 1/1
was silently dropped, effectively halting the state machine, if the
state machine was still believed to be on 1/0. That was wrong behavior,
of course.

The updated TestTx_multiple_simpleSend test still isn't passing, but now
we are correctly replaying the first header to the state machine. At the
end of the test, the state machine is still reported to be at 1/1, even
though we have replayed later headers. Now that I have a better handle
on what is wrong here, debugging the next issue ought to be faster than
this one was.