feat: add tmsqlite module
This will move out to its own module following open sourcing of the core
repository; they are together for now, to avoid the pain of managing
private Go modules.

The mirror store is implemented first here, and the compliance tests
were backfilled.

I was going to do the action store next, but it has a Load method which
needs to be renamed so that we can clearly satisfy multiple interfaces
with that single TMStore type.

I am not yet sure about the multiple ways validators are currently
stored in the tmstore types. Maybe we will push them all into the
ValidatorStore, or maybe we will store the validator hashes only, or
maybe the usage will change altogether.