refactor: use ValidatorSet values in FinalizationStore
The flat list of validators seemed incorrect here, and it isn't that
much more effort to deal with a proper ValidatorSet.

In updating this code, it looks like it might be worth adding a way to
retrieve only the validator set from a finalized height, but we will
save that for another time.