feat: address remaining compliance issues for action store
The last one was a particular error if different public keys were
recorded in the action store across prevotes and precommits.

There are further tests that should be added to the action store
compliance tests, but I am going to move on to the remaining round store
so we can start integration before further hardening the stores.