chore: add height column to proposed_headers table
This allows us to add a unique constraint on height, round, key so that
we can trivially catch attempts to add a second proposed block from the
same proposer in the same round. It probably simplifies lookups too,
since those would usually be by height and round; then we don't have to
join on the headers table to perform the query.

Also stub out the start of the RoundStore implementation.