refactor: add IsValidKeyID to CommonMessageSignatureProofScheme
This circumvents tracking a dedicated nil proof for key ID verification.
It seems reasonable to push this reponsibility to the scheme.
Non-aggregating schemes will have some way of representing the ID by a
singular index, while aggregating schemes will likely do some kind of
verification that the ID is following scheme-specific aggregation rules,
such as needing to fill leftmost first, needing to fill pairwise in a
binary tree shape, or so on.

Next up is removing all notions of the nil proof being required.