chore: suppress misleading unique constraint error
We expect to not save the same proposed header twice, but in the case of
tmsqlite, the action store and the round store have a shared header
table that will save successfully on the state machine action, but then
reach a unique constraint violation when the round store tries to save
identical data. We could special case the saving pattern, but then we
would have to teach the kernel more about the store; it seems better to
just react to the error, at least for now.