fix: only set committed header flag on committed headers
The tmsqlite store was incorrectly setting committed=true on proposed
headers, which was causing the committed header lookup to fail, since
there was not an actual commit proof to go along with the proposed
headers.

We have now backfilled tests for this in tmsqlite, but those tests do
need to move somewhere into gordian/tm/tmstore so that they can be
shared with the next persistent store.

Also create a read-write and a read-only database connection for
tmsqlite, and use separate constructor functions for an in-memory
database vs. an on-disk database. (This should have been its own commit,
but that's the way it goes sometimes on a bug hunt like this one.)