chore: add replayed header concept to round store
The replayed headers act a little like plain proposed headers, but they
are different enough that it seems valid to add them to the round store
as their own methods.

Tests are still passing with gcosmos using tmmemstore. I think there are
more edge cases to shake out with tmsqlite.