feat: enable tmsqlite+gcosmos
It is hardcoded to use an on-disk database (at a temporary file
location). We need to get some command-line flags in place to allow
toggling between tmsqlite in-memory, tmsqlite on-disk, and the original
tmmemstore implementations.

This change also fixes outstanding flakiness in the tmsqlite in-memory
store, which seemed to be due to allowing more than one concurrent
write.

Also expand the inline documentation, in case we ever need to touch any
of the SQLite setup again in the future.