feat: use command line flag to choose consensus database
This is all behind the --g-sqlite-path flag, which could possibly be
renamed later.

If the value is empty, use the tmmemstore stores.

If the value is the exact string :memory:, which already holds special
meaning in SQLite, use the in-memory tmsqlite database.

Otherwise the value is treated as the path to an SQLite database, which
is permitted to not yet exist.

To control which of these behaviors to use in the gcosmos tests, edit
the useMemStore and useSQLiteInMem constants at the top of main_test.go.