chore: fully close root store on gordian shutdown
And halfway implement the test.

There were multiple parts to this. First, we include a new SDK patch
that closes the underlying "raw database" when the root store is closed.
(The patch also includes expanded logging on server component start and
stop.)

Then we have to actually call store.Close in the Gordian component's
Stop method.

Next, we have a new (but currently skipped) test that starts a single
gcosmos validator, stops, it, and starts it again.

I am seeing three different outcomes in the test currently.

1. We still get a "resource temporarily unavailable" error as though we
   didn't shut down the underlying database. I don't know the root cause
   on why this sometimes happens. The logs appear as though we aren't
   stopping, iirc. It is possibly just a test timing issue.
2. Sometimes we get a connection refused (I think) error when retrieving
   the watermark from the restarted validator. I suspect this is the
   test reading a stale value from the previous run; we might need to
   wipe out the old HTTP address file before attempting to restart.
3. Occasionally we appear to actually restart the validator, and then we
   get a runtime panic about attempting to hash a set of zero
   validators. This is exactly the kind of error this set of tests was
   intended to discover.

The second case is probably the simplest to fix, as the first case
likely involves another SDK patch, and the third case will involve
backfilling further store tests and possibly some engine
integration-style tests.