fix: use state machine store from state machine
This looks like it enables stops and restarts for the state machine,
based on some debug logs; but the gcosmos integration test is still not
yet passing. Besides the rare database resource temporarily unavailable
error, it is consistently failing with connection refused on the new
HTTP server.

As far as I can see, the address file is distinct and new from the
original run. It is most likely a test setup issue, but we should be
able to get to the bottom of it either way tomorrow.