test: run only the required validators
We previously ran into an undocumented bug where validator counts were
incorrect when above 1 or below 11, so we ran 4 validators and padded
them with 7 "dummy" validators. The dummy validators were given minimal
stake so as to not be required for consensus. This had the unfortunate
downside of wasting rounds when it was their turn to validate.

We successfully ignored (or at least worked around) the problem until it
went away. Running just 4 validators in our multi-validator tests works
as intended, now.