build: add skeleton Makefile
This is my preferred pattern for help -- being the default target, and
being embedded in the Makefile. Automated "make help" output is a poor
compromise in my experience.

And this establishes the style for .PHONY; includes examples of @ to
suppress a command being printed; and also has an example of dependency
between tasks.